third_party/WebKit/LayoutTests/crypto/rsa-ssa-generateKey-parseAlgorithm-failures.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Tests generateKey() with bad RSASSA-PKCS1-v1_5 parameters.");
jsTestIsAsync = true;

extractable = true;
keyUsages = ['sign', 'verify'];

Promise.resolve(null).then(function() {
    return crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5', hash: {name: 'sha-1'}, modulusLength: -30}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5', hash: {name: 'sha-1'}, modulusLength: NaN}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5', hash: {name: 'sha-1'}}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5', hash: {name: 'sha-1'}, modulusLength: 10}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5', hash: {name: 'sha-1'}, modulusLength: 10, publicExponent: 10}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5', hash: {name: 'sha-1'}, modulusLength: 10, publicExponent: null}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5', hash: {name: 'sha-1'}, modulusLength: 10, publicExponent: new Uint8Array(0)}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5', hash: {name: 'sha-1'}, modulusLength: 1024, publicExponent: new Uint8Array(0)}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5', hash: {name: 'NOT-A-HASH'}, modulusLength: 10, publicExponent: new Uint8Array(3)}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    return crypto.subtle.generateKey({name: 'RSASSA-PKCS1-v1_5', modulusLength: 10, publicExponent: new Uint8Array(3)}, extractable , keyUsages);
}).then(failAndFinishJSTest, function(result) {
    logError(result);
}).then(finishJSTest, failAndFinishJSTest);

</script>

</body>
</html>
