third_party/WebKit/LayoutTests/crypto/rsa-ssa-generateKey-parseAlgorithm-failures-expected.txt

Tests generateKey() with bad RSASSA-PKCS1-v1_5 parameters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


error is: SyntaxError: RsaHashedKeyGenParams: modulusLength: Outside of numeric range
error is: SyntaxError: RsaHashedKeyGenParams: modulusLength: Is not a number
error is: SyntaxError: RsaHashedKeyGenParams: modulusLength: Missing required property
error is: SyntaxError: RsaHashedKeyGenParams: publicExponent: Missing or not a Uint8Array
error is: SyntaxError: RsaHashedKeyGenParams: publicExponent: Missing or not a Uint8Array
error is: SyntaxError: RsaHashedKeyGenParams: publicExponent: Missing or not a Uint8Array
error is: NotSupportedError: The modulus length must be a multiple of 8 bits and >= 256 and <= 16384
error is: DataError: The "publicExponent" must be either 3 or 65537
error is: NotSupportedError: RsaHashedKeyGenParams: hash: Algorithm: Unrecognized name
error is: SyntaxError: RsaHashedKeyGenParams: hash: Missing or not a dictionary
PASS successfullyParsed is true

TEST COMPLETE
